More aggressive sync behaviour
The previous implementation let the OS keep store data in writeback
cache, this is nice for performance but is not good for integrity.

We now say that we want to avoid this, naturally this comes with as much
guarentee as the OS provides (read near none) but its the best we get.